==> 00-West-Nile-clean-stage-1.ipynb <==
# West Nile Project

"Given weather, location, testing, and spraying data, this competition asks you to predict **when and where different species of mosquitos will test positive for West Nile virus**. A more accurate method of predicting outbreaks of West Nile virus in mosquitos will help the City of Chicago and CPHD more efficiently and effectively allocate resources towards preventing transmission of this potentially deadly virus." kaggle

## Import Libraries


#imports
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy.stats as stats

#additional libraries
import datetime as dt

## Import the Data

#long & lat for places in chi
spray = pd.read_csv('./data/spray.csv')

#weather data
weather = pd.read_csv('./data/weather.csv')

#train & test sets
test = pd.read_csv('./data/test.csv')
train = pd.read_csv('./data/train.csv')

## Cleaning Test & Train

#### Spray

spray.info()

spray.head()

spray.Date = pd.to_datetime(spray.Date)

#since only time has NaNs, taking a shortcut to fill all NaNs with 12PM
#Note that we can't use time to measure effectiveness of spray now, if that ever comes up


spray.fillna('12:00:00 PM', inplace=True)

#did not use this cell as it was adding the current date to the time
#spray.Time = pd.to_datetime(spray.Time)


spray.info()

#10 sprays total
spray.Date.value_counts()

spray.set_index('Date', inplace=True)

spray_centers = spray.groupby('Date').mean()

#Because this is showing the mean lat/long, I think it's safe to call this the center spray point.  
#It may be useful to get the boundaries of the spray too though

spray_centers

#These could come in handy if we wanted to map the area covered by each spray later 
#(like the states were mapped in class, or on tableau)

min_latitude = []
max_latitude = []
min_longitude = []
max_longitude = []

for index in spray_centers.index:
    min_latitude.append(spray[(spray.index == index)].Latitude.min())
    max_latitude.append(spray[(spray.index == index)].Latitude.max())
    min_longitude.append(spray[(spray.index == index)].Longitude.min())
    max_longitude.append(spray[(spray.index == index)].Longitude.max())
    

spray_centers['min_latitude'] = min_latitude
spray_centers['max_latitude'] = max_latitude
spray_centers['min_longitude'] = min_longitude
spray_centers['max_longitude'] = max_longitude

spray_centers

#Some days, spraying was concentrated in one area, other days it spread in one or both directions

#This was cool, even though the value_counts are sorted, they lined up to the proper indexes.

spray_centers['Total_sprays'] = spray.index.value_counts()

spray_centers

#### Train/Test

'train.csv, test.csv - the training and test set of the main dataset. The training set consists of data from 2007, 2009, 2011, and 2013, while in the test set you are requested to predict the test results for 2008, 2010, 2012, and 2014.' - kaggle

train.head(1)

test.head(1)

##### adding zipcodes.

## Looked through the addresses, and many do not have a zip.
## Going to loop through and see what kind of values come back

zipcodes = []
for add in train.Address:
    zipcodes.append(add[-10:-5])

#returned 51 zip codes, some are missing zips though

len(set(zipcodes))

#14.3 percent of zipcodes will be missing.  I think we can still dummy them out and 
#get value in the model, but I'll hold off on dummies for now.

zipcodes.count('o, IL')/len(zipcodes)

zipcodes = ['Unk' if x == 'o, IL' else x for x in zipcodes]

train.insert(loc=2, column='zipcode', value=zipcodes)

train.head(1)

#repeating for test

zipcodes2 = []
for add in test.Address:
    zipcodes2.append(add[-10:-5])

#returned 51 zip codes, some are missing zips though.
#hopefully they're the same 51 zips for when we dummy.  we'll cross that bridge later.

len(set(zipcodes2))

zipcodes2.count('o, IL')/len(zipcodes2)

zipcodes2 = ['Unk' if x == 'o, IL' else x for x in zipcodes2]

test.insert(loc=2, column='zipcode', value=zipcodes2)

test.head(1)

train.drop(['Address', 'Street', 'AddressNumberAndStreet'], axis=1, inplace=True)

train.head(2)

test.drop(['Address', 'Street', 'AddressNumberAndStreet'], axis=1, inplace=True)

test.head(2)

train = pd.get_dummies(train, columns=['Species', 'zipcode', 'Trap']) 

test = pd.get_dummies(test, columns=['Species', 'zipcode', 'Trap']) 

train.shape

test.shape

(set(test.columns)) - (set(train.columns))

train['Species_UNSPECIFIED CULEX'] = [0] * train.shape[0]
train['Trap_T002A'] = [0] * train.shape[0]
train['Trap_T002B'] = [0] * train.shape[0]
train['Trap_T065A'] = [0] * train.shape[0]
train['Trap_T090A'] = [0] * train.shape[0]
train['Trap_T090B'] = [0] * train.shape[0]
train['Trap_T090C'] = [0] * train.shape[0]
train['Trap_T128A'] = [0] * train.shape[0]
train['Trap_T200A'] = [0] * train.shape[0]
train['Trap_T200B'] = [0] * train.shape[0]
train['Trap_T218A'] = [0] * train.shape[0]
train['Trap_T218B'] = [0] * train.shape[0]
train['Trap_T218C'] = [0] * train.shape[0]
train['Trap_T234'] = [0] * train.shape[0]

train.shape

test.shape

(set(test.columns)) - (set(train.columns))

(set(train.columns)) - (set(test.columns))

train.to_csv('west_nile_train.csv')

test.to_csv('west_nile_test.csv')

^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
train test cleaned, dummied
---------


___________

---------

anything under here is unstructured/undone/meaningless

-----------

len(train.Trap.value_counts())
#len(train.Trap.unique()) samesame

len(test.Trap.value_counts())

train.sort_values(by='NumMosquitos', ascending = False).head()

#Species breakdown when WNV present
#May not be strong enough because of low count of other species.

train[(train.WnvPresent == 1)].Species.value_counts()

#total Species breakdown wnv

train.Species.value_counts()

#test.Species is balanced

test.Species.value_counts()

len(train.Block.value_counts())

len(test.Block.value_counts())

len(weather.Date.unique())
