==> 04-spray-train-weather-concat.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
pd.set_option("display.max_columns",500)
pd.set_option("display.max_rows",500)

import matplotlib.pyplot as plt

train = pd.read_csv('./data/west_nile_train_spray.csv', index_col='Date')
test = pd.read_csv('./data/west_nile_test_spray.csv', index_col='Date')
weather_1 = pd.read_csv('./data/weather-data/weather_1_UNSCALED.csv', index_col='Date')
weather_2 = pd.read_csv('./data/weather-data/weather_2_UNSCALED.csv', index_col='Date')

# Convert all dfs to datetime

weather = pd.concat([weather_1, weather_2])
train.drop('Unnamed: 0', axis=1, inplace=True)
test.drop('Unnamed: 0', axis=1, inplace=True)

#make this a datetime measure
weather.index = pd.to_datetime(weather.index)
train.index = pd.to_datetime(train.index)
test.index = pd.to_datetime(test.index)

# Aggregate weather station data

#merge weather stations on date
weather = weather.groupby(weather.index).agg('mean')

#export a copy of the aggregated weather station data
#weather.to_csv('./data/weather-data/01_weather_concat.csv')

# Merged weather into train and test data

df = pd.merge(train, weather, left_index=True, right_index=True)
df_test = pd.merge(test, weather, left_index=True, right_index=True)

#weather to csv
#df.to_csv('./02_weather_spray_train.csv')

#df_test to csv
#df_test.to_csv('./02_weather_spray_test.csv')

# Checked shape of data

#train has 25 dates
train[train.index == '2007-05-29'].shape

#weather has 1 date
weather[weather.index == '2007-05-29'].shape

#new df has 25 dates with the same weather info
df[df.index == '2007-05-29'].shape

#classes are still balanced
df.WnvPresent.value_counts(dropna=False)
