==> 07-Balanced-Model-All-Features-PCA.ipynb <==
# Import Libraries

import pandas as pd
pd.set_option("display.max_columns",500)
pd.set_option("display.max_rows",500)

import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.tree import DecisionTreeClassifier, ExtraTreeClassifier
from sklearn.ensemble import BaggingClassifier, RandomForestClassifier, VotingClassifier, GradientBoostingClassifier, AdaBoostClassifier
from sklearn import svm

from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

# Import Train/Test Data

#import data
train = pd.read_csv('./west-nile-project-4/data/weather-data/02_weather_spray_train.csv', index_col=0)
test = pd.read_csv('./west-nile-project-4/data/weather-data/02_weather_spray_test.csv', index_col='Id')

train.WnvPresent.value_counts()

# Set X and y variables

#id X and y
X = train.drop(['NumMosquitos', 'WnvPresent', 'Latitude', 'Longitude', 'Sprays_In_Last_Month'], axis=1)
y = train.WnvPresent

# Train Test Split

#train split
X_train, X_test, y_train, y_test = train_test_split(X, y)

# Standard Scaler

ss = StandardScaler()
X_train = ss.fit_transform(X_train)
X_test = ss.transform(X_test)

# PCA

pca = PCA()

Z_train = pca.fit_transform(X_train)
Z_test = pca.transform(X_test)
Z_train.shape

# Cross Val Score Potential Classification Models

#cross val score all potential classification models
lr = LogisticRegression() 
knn = KNeighborsClassifier() 
#nb = MultinomialNB() won't take neg vals
dt = DecisionTreeClassifier() 
et = ExtraTreeClassifier()
bag = BaggingClassifier()
rf = RandomForestClassifier()
gb = GradientBoostingClassifier()
ad = AdaBoostClassifier()
#svm1 = svm.SVC() #too slow, not strong performer


print('lr', cross_val_score(lr, Z_train, y_train).mean())
print('knn', cross_val_score(knn, Z_train, y_train).mean())
#print('nb', cross_val_score(nb, X_train, y_train).mean()) 
print('dt', cross_val_score(dt, Z_train, y_train).mean()) 
print('et', cross_val_score(et, Z_train, y_train).mean()) 
print('bag', cross_val_score(bag, Z_train, y_train).mean())
print('rf', cross_val_score(rf, Z_train, y_train).mean()) 
print('gb', cross_val_score(gb, Z_train, y_train).mean()) 
print('ad', cross_val_score(ad, Z_train, y_train).mean()) 
#print('svm', cross_val_score(svm1, X_train, y_train).mean()) #too slow, not strong performer

# Making an Interpretable Equation

# ROC AUC Score

#printing the model coefs
lr.fit(Z_train, y_train)
lr_coefs = pd.DataFrame(lr.coef_, columns=X.columns, index=['coef']).T
lr_coefs.coef.sort_values(ascending=False)
#higher coefs are more likely to indicate WNV

#roc auc score
lr_proba = lr.predict_proba(Z_test)
roc_auc_score(y_test, lr_proba[:, 1])

# Random Forest Model

rf.fit(Z_train, y_train)
#rf.score(X_test, y_test)

rf_proba = rf.predict_proba(Z_test)
roc_auc_score(y_test, rf_proba[:, 1])

# Gradient Boost

gb.fit(Z_train, y_train)
gb_proba = gb.predict_proba(Z_test)
roc_auc_score(y_test, gb_proba[:, 1])

# Prepping for Export

test_ss = ss.transform(test.drop(['Date', 'Latitude', 'Longitude', 'Sprays_In_Last_Month'], axis=1))
test_pca = pca.transform(test_ss)

gb_preds = gb.predict(test_pca)
gb_preds

#exporting to csv
pd.DataFrame({'WnvPresent':gb_preds}, index=test.index).to_csv('./preds5.csv')
