==> 09-Model-Gridsearch.ipynb <==
# Import Libraries

import pandas as pd
pd.set_option("display.max_columns",500)
pd.set_option("display.max_rows",500)

from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import accuracy_score
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.tree import DecisionTreeClassifier, ExtraTreeClassifier
from sklearn.ensemble import BaggingClassifier, RandomForestClassifier, VotingClassifier, GradientBoostingClassifier, AdaBoostClassifier
from sklearn import svm

# Basic Model
## Import Train/Test Data

#import data
test = pd.read_csv('./data/west_nile_test.csv', index_col=0)
train = pd.read_csv('./data/west_nile_train.csv', index_col=0)

## Set X and y variables

#id X and y
X = train.drop(['Date', 'NumMosquitos', 'WnvPresent'], axis=1)
y = train.WnvPresent

## Train Test Split

#train split
X_train, X_test, y_train, y_test = train_test_split(X, y)

## Cross Val Score Potential Classification Models

#cross val score all potential classification models
lr = LogisticRegression() 
knn = KNeighborsClassifier() 
#nb = MultinomialNB() won't take neg vals
dt = DecisionTreeClassifier() 
et = ExtraTreeClassifier()
bag = BaggingClassifier()
rf = RandomForestClassifier()
gb = GradientBoostingClassifier()
ad = AdaBoostClassifier()
svm1 = svm.SVC()


print('lr', cross_val_score(lr, X_train, y_train).mean())
print('knn', cross_val_score(knn, X_train, y_train).mean())
#print('nb', cross_val_score(nb, X_train, y_train).mean()) 
print('dt', cross_val_score(dt, X_train, y_train).mean()) 
print('et', cross_val_score(et, X_train, y_train).mean()) 
print('bag', cross_val_score(bag, X_train, y_train).mean())
print('rf', cross_val_score(rf, X_train, y_train).mean()) 
print('gb', cross_val_score(gb, X_train, y_train).mean()) 
print('ad', cross_val_score(ad, X_train, y_train).mean()) 
print('svm', cross_val_score(svm1, X_train, y_train).mean())

- All models seem to reveal similar enough results. 
- LogReg choosen for ease of interpretability.

#fit and score
lr.fit(X_train, y_train)
lr.score(X_test, y_test)

lr_coefs = pd.DataFrame(lr.coef_, columns=X.columns, index=['coef']).T
lr_coefs.coef.sort_values(ascending=False)
#higher coefs are more likely to indicate WNV

# Balanced Model
## Import Data

#import data
balenced_train = pd.read_csv('./data/BALANCED_west_nile_train.csv', index_col=0)

balenced_train.WnvPresent.value_counts()

# Set X and y variables

#id X and y
X = balenced_train.drop(['Date', 'NumMosquitos', 'WnvPresent', 'Latitude', 'Longitude'], axis=1)
y = balenced_train.WnvPresent

# Train Test Split

#train split
X_train, X_test, y_train, y_test = train_test_split(X, y)

# Cross Val Score Potential Classification Models

#cross val score all potential classification models
lr = LogisticRegression() 
knn = KNeighborsClassifier() 
#nb = MultinomialNB() won't take neg vals
dt = DecisionTreeClassifier() 
et = ExtraTreeClassifier()
bag = BaggingClassifier()
rf = RandomForestClassifier()
gb = GradientBoostingClassifier()
ad = AdaBoostClassifier()
#svm1 = svm.SVC() #takes a long time, doesn't score partic well.


print('lr', cross_val_score(lr, X_train, y_train).mean())
print('knn', cross_val_score(knn, X_train, y_train).mean())
#print('nb', cross_val_score(nb, X_train, y_train).mean()) 
print('dt', cross_val_score(dt, X_train, y_train).mean()) 
print('et', cross_val_score(et, X_train, y_train).mean()) 
print('bag', cross_val_score(bag, X_train, y_train).mean())
print('rf', cross_val_score(rf, X_train, y_train).mean()) 
print('gb', cross_val_score(gb, X_train, y_train).mean()) 
print('ad', cross_val_score(ad, X_train, y_train).mean()) 
#print('svm', cross_val_score(svm1, X_train, y_train).mean()) #.644


lr.fit(X_train, y_train)
lr_coefs = pd.DataFrame(lr.coef_, columns=X.columns, index=['coef']).T
lr_coefs.coef.sort_values(ascending=False)
#higher coefs are more likely to indicate WNV

## Gridsearch

from sklearn.model_selection import GridSearchCV
from sklearn.metrics import recall_score, make_scorer, f1_score

#Logistic Regression
grid_params = {
    'max_iter': [50,100,300,1000],
    'class_weight':['balanced'],
    'solver':['newton-cg','lbfgs','liblinear','sag','saga'],
    'multi_class':['ovr','multinomial']
    
}

gs = GridSearchCV(
    LogisticRegression(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#knn
grid_params = {
    'n_neighbors':[3,5,7],
    'weights':['uniform','distance'],
    'metric':['minkowski']
}

gs = GridSearchCV(
    KNeighborsClassifier(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#dt
grid_params = {
    'criterion':['gini','entropy'],
    'splitter':['best','random'],
    'min_samples_leaf':[1,2,4],
    'max_depth':[50,100,115,200,300],
    'min_samples_split':[2,3,5]
}

gs = GridSearchCV(
    DecisionTreeClassifier(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#et
grid_params = {
    'criterion':['gini','entropy'],
    'splitter':['best','random'],
    'min_samples_leaf':[1,2,4,5,8,],
    'max_depth':[50,100,300,500],
    'min_samples_split':[2,3,5]
    
}

gs = GridSearchCV(
    ExtraTreeClassifier(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#bag
grid_params = {
    'n_estimators':[5,10,13],
    'max_samples':[3,5,10,50,100],
    #'max_features':[0,2,3],
    'bootstrap_features':['False','True'],
    
}

gs = GridSearchCV(
    BaggingClassifier(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#rf
grid_params = {
    'criterion':['gini','entropy'],
    'n_estimators':[5,10,15],
    'min_samples_leaf':[1,2,4,5,8,],
    'max_depth':[50,100,300,500],
    'min_samples_split':[2,3,5]
    
}

gs = GridSearchCV(
    RandomForestClassifier(),
    grid_params,
    verbose = 1,
    cv = 5,
    n_jobs = -1
)

gs_results = gs.fit(X_train,y_train)

print('best param',gs.best_params_)
print('best score',gs.best_score_)

#gb
#hadn't looked at this because nothing has had a better score

#ad
##hadn't looked at this because nothing has had a better score
